Decommission strong unit conversions
